% tinyrox says don't edit this manually, but it can't stop you!
\name{add_history_entry}
\alias{add_history_entry}
\title{Add entry to history}
\usage{
add_history_entry(history, entry)
}
\arguments{
\item{history}{Current history list.}

\item{entry}{New entry to add.}
}
\description{
Add entry to history
}
\keyword{internal}
