% tinyrox says don't edit this manually, but it can't stop you!
\name{create_history_entry}
\alias{create_history_entry}
\title{Create a new history entry}
\usage{
create_history_entry(text, voice, backend, model = NULL, audio_file = NULL,
                     params = NULL)
}
\arguments{
\item{text}{Input text.}

\item{voice}{Voice used.}

\item{backend}{Backend used.}

\item{model}{Model used (if any).}

\item{audio_file}{Path to audio file (optional).}
}
\description{
Create a new history entry
}
\keyword{internal}
