% tinyrox says don't edit this manually, but it can't stop you!
\name{delete_history_entry}
\alias{delete_history_entry}
\title{Delete history entry}
\usage{
delete_history_entry(history, id)
}
\arguments{
\item{history}{Current history list.}

\item{id}{Entry ID to delete.}
}
\description{
Delete history entry
}
\keyword{internal}
