% tinyrox says don't edit this manually, but it can't stop you!
\name{format_timestamp}
\alias{format_timestamp}
\title{Format timestamp for display}
\usage{
format_timestamp(timestamp)
}
\arguments{
\item{timestamp}{POSIXct timestamp.}
}
\description{
Format timestamp for display
}
\keyword{internal}
