% tinyrox says don't edit this manually, but it can't stop you!
\name{icon_button}
\alias{icon_button}
\title{A button whose label is an icon}
\usage{
icon_button(id, name, title, class = NULL)
}
\arguments{
\item{id}{character input ID}

\item{name}{character icon name}

\item{title}{character tooltip and accessible label}

\item{class}{character extra CSS class(es)}
}
\value{
A glinty UI element
}
\description{
A button whose label is an icon
}
\keyword{internal}
