% tinyrox says don't edit this manually, but it can't stop you!
\name{icon_label_button}
\alias{icon_label_button}
\title{A button with an icon beside its label}
\usage{
icon_label_button(id, name, label, class = NULL)
}
\arguments{
\item{id}{character input ID}

\item{name}{character icon name}

\item{label}{character button text}

\item{class}{character extra CSS class(es)}
}
\value{
A glinty UI element
}
\description{
A button with an icon beside its label
}
\keyword{internal}
