% tinyrox says don't edit this manually, but it can't stop you!
\name{icon_shapes}
\alias{icon_shapes}
\title{Child elements for one icon}
\usage{
icon_shapes(name)
}
\arguments{
\item{name}{character icon name}
}
\value{
list of glinty UI elements
}
\description{
Child elements for one icon
}
\keyword{internal}
