% tinyrox says don't edit this manually, but it can't stop you!
\name{save_audio_file}
\alias{save_audio_file}
\title{Save audio file to history}
\usage{
save_audio_file(audio_data, entry_id, format = "wav")
}
\arguments{
\item{audio_data}{Raw audio bytes.}

\item{entry_id}{History entry ID.}

\item{format}{Audio format (wav, mp3).}
}
\description{
Save audio file to history
}
\keyword{internal}
