% tinyrox says don't edit this manually, but it can't stop you!
\name{truncate_text}
\alias{truncate_text}
\title{Truncate text for preview}
\usage{
truncate_text(text, max_length = 50)
}
\arguments{
\item{text}{Text to truncate.}

\item{max_length}{Maximum length.}
}
\description{
Truncate text for preview
}
\keyword{internal}
